build: add commitlint, update docs to point to Yotta instead of pmoeller91-yottaa
This change adds commitlint to best support the release-it action, by adding a commit-msg hook
locally as well as a check as part of the test step in the Test GitHub action. Documentation is
updated to point universally to Yottaa instead of specifically to pmoeller91-yottaa to reflect the
package ownership transfer.